partner page done
